+++ _examples/01_minimal/main.go
@@ -1,0 +1,18 @@
+package main
+
+import "github.com/go-joe/joe"
+
+func main() {
+	b := joe.New("example-bot")
+	b.Respond("ping", Pong)
+
+	err := b.Run()
+	if err != nil {
+		b.Logger.Fatal(err.Error())
+	}
+}
+
+func Pong(msg joe.Message) error {
+	msg.Respond("PONG")
+	return nil
+}